Updating the interfaces for working with the web
