--- argos/datalayer/__init__.py
@@ -1,1 +1,0 @@
-from .argosWebDatalayer import GQLDataLayerFactory
