+++ argos/thingsboard/tbHome.py
@@ -1,14 +1,11 @@
 import os
 import json
 import requests
+
+from argos.thingsboard.tbEntitiesProxy import tbEntityHome
 from .tb_api_client.swagger_client import ApiClient, Configuration
-from .tb_api_client.swagger_client import Asset, Device
-from .tb_api_client.swagger_client.rest import ApiException
 from .tb_api_client.swagger_client import DeviceControllerApi, AssetControllerApi, EntityRelationControllerApi, AlarmControllerApi
 from .tb_api_client.swagger_client import TelemetryControllerApi
-
-
-from .tbEntitiesProxy import DeviceProxy, AssetProxy
 
 tojson = lambda x: json.loads(
     str(x).replace("None", "'None'").replace("'", '"').replace("True", "true").replace("False", "false"))
@@ -248,171 +245,3 @@ class swaggerAPI(object):
         return api_client
 
 
-class tbEntityHome(dict):
-    """
-        This class manages the creation, removal and retrieving of an entity (ASSET or DEVICE for now) from the server.
-        The design pattern is a thin proxy. so that the proxy does not cache. Thincker clients can be
-        implemented in the future if necessary.
-
-        Inherits from a dictionary and holds the current instances of the entity proxy.
-    """
-    _swagger = None
-
-    _entityType = None
-    _entityApi = None
-    _entityProxy = None
-
-    def __init__(self, iswaggerApi, entityType):
-        """
-            Initializes the Home.
-
-        :param iswaggerApi:
-                the swagger wrapper with all the api initialized and after authentication with the server.
-        :param entityType:
-                asset or device.
-        """
-        proxyClass = {"asset": AssetProxy, "device": DeviceProxy}
-        objClass = {"asset": Asset, "device": Device}
-
-        self._entityType = entityType.lower()
-        self._swagger = iswaggerApi
-        self._entityApi = iswaggerApi.getApi(self._entityType)
-        self._entityProxy = proxyClass[self._entityType]
-        self._entityObj = objClass[self._entityType]
-
-    def createProxy(self, entityName, entityType=None):
-        """
-            Creates a proxy entity.
-
-            Call modes:
-
-            createProxy(entityName = "NewDevice")
-
-                    Create a proxy for the entity NewDevice.
-                    if does not exist raise exception.
-
-            createProxy(entityName = "NewDevice",entityType="blah")
-                    Create a proxy for the entity NewDevice.
-                    if does not exist - create.
-                    if exists and type mismatch the entityType - raise exception.
-
-            :return returns the proxy object of the entity according to the home type.
-        """
-
-        entityData = self.getEntity(entityName)
-
-        if entityType is None:
-            # kwargs is not emtpy.
-            if entityData is not None:
-                newEntityProxy = self._entityProxy(entityData, swagger=self._swagger, home=self)
-            else:
-                raise ValueError("Device %s does not exist" % entityName)
-
-        else:
-            if entityData is None:
-                newentity = self._entityObj(name=entityName, type=entityType)
-                saveFunc = getattr(self._entityApi, "save_%s_using_post" % self._entityType)
-                saveFunc(newentity)
-                newEntityProxy = self._entityProxy(self.getEntity(entityName), swagger=self._swagger, home=self)
-            else:
-                if entityData["type"] != entityType:
-                    raise ValueError(
-                        "Cannot create Proxy. The type of %s mismatch. The device type in Thingsboard is %s while requested type is %s " % (
-                        entityName, entityData["type"], entityType))
-                newEntityProxy = self._entityProxy(entityData, swagger=self._swagger, home=self)
-
-        self[entityName] = newEntityProxy
-        return newEntityProxy
-
-    def getEntity(self, entityName):
-        """
-            Gets the device data from the TB server.
-
-
-        :param entityName: The name of the device.
-
-        :return:
-                return a dict:
-
-                .. code-block:: json
-
-                {
-                 'additional_info': <..>,
-                 'created_time': <timestamp>,
-                 'customer_id': {'id': <str> },
-                 'id': {'id': <str>},
-                 'name': <str>,
-                 'tenant_id': {'id': <str>},
-                 'type': <str>
-                }
-
-                if device exists, False otherwise.
-        """
-        ret = None
-        try:
-            getFunc = getattr(self._entityApi, "get_tenant_%s_using_get_with_http_info" % self._entityType)
-            data, _, _ = getFunc(entityName)
-            ret = data.to_dict()
-        except ApiException as e:
-            if json.loads(e.body)['errorCode'] != 32:
-                raise e
-
-        return ret
-
-    def exists(self, entityName):
-        """
-            Checks if the device exists in the TB server.
-
-        :param entityName: The name of the device.
-
-        :return: True if device exists, false otherwise.
-        """
-        return False if self.get(entityName) is None else True
-
-    def delete(self, entityName):
-        try:
-
-            deleteFunc = getattr(self._entityApi, "delete_%s_using_delete" % self._entityType)
-            deleteFunc(self[entityName].id)
-            del self[entityName]
-        except ApiException as e:
-            pass
-
-    def getAllEntitiesName(self, limit=10000):
-        ret = None
-        try:
-            getFunc = getattr(self._entityApi, "get_tenant_%ss_using_get_with_http_info" % self._entityType)
-            data, _, _ = getFunc(limit)
-            ret = data.to_dict()['data']
-            #ret = list(map(lambda x: x['name'], ret))
-            ret = [x['name'] for x in ret]
-        except ApiException as e:
-            if json.loads(e.body)['errorCode'] != 32:
-                raise e
-
-        return ret
-
-    def getAllEntitiesNameByType(self, type, limit=10000):
-        ret = None
-        try:
-            getFunc = getattr(self._entityApi, "get_tenant_%ss_using_get_with_http_info" % self._entityType)
-            data, _, _ = getFunc(limit)
-            ret = data.to_dict()['data']
-            ret = [x['name'] for x in ret if x['type']==type]
-        except ApiException as e:
-            if json.loads(e.body)['errorCode'] != 32:
-                raise e
-
-        return ret
-
-    def deleteAllEntities(self):
-        entitiesName = self.getAllEntitiesName(limit=10000)
-        for entityName in entitiesName:
-            self.delete(entityName)
-
-    def __getitem__(self, item):
-        if item not in self:
-            ret = self.createProxy(item)
-        else:
-            ret = self.get(item)
-        return ret
